requirements.txt
--extra-index-url=https://MLibSpotify:oqjfcg6dxemx66z2mnxmi6yfcrm7a62n553hcdu47uqmopjx6pza@pkgs.dev.azure.com/MikeKennedyDev/Libraries/_packaging/MLibSpotify/pypi/simple/
MLibSpotify==2.1.18
discord.py==2.0.1
python-dotenv~=0.21.0
pyodbc~=4.0.34